pyproject.toml
[project]
name = "oilfield-api-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-ninja>=1.5.3",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
]
